% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{class}
\name{Benneat-class}
\alias{Benneat-class}
\title{Benneat}
\description{
A class used to store the benchmark dataset with the same length so called neat.
}
\section{Slots}{

\describe{
\item{\code{Seqs}}{A vector contains the sequences}

\item{\code{labs}}{A vector contains the labels like "methylated"/"unmethylated"}

\item{\code{seq_num}}{A numeric denotes the num of seqlist}

\item{\code{lab_num}}{A table contains the num of each Label}

\item{\code{seq_length}}{A numerci denotes the length of each sequence}
}}

