% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNACGR.R
\name{DNACGR}
\alias{DNACGR}
\title{DNACGR}
\usage{
DNACGR(Seq, Xdivide = 4, Ydivide = 4)
}
\arguments{
\item{Seq}{String. Which denotes the DNA sequence}

\item{Xdivide}{numeric. denotes the number of X to be splitted. default:4}

\item{Ydivide}{numeric. denotes the number of Y to be splitted. default:4}
}
\value{
A number which is the fscore

a numeric vector with length of Xdivide*Ydivide
}
\description{
The formula is from J.Jia's JTB paper about PPI in 2019. In the
square, A=(1,1),T=(0,1),C=(0,0),G=(1,0)
}
\details{
The function implement the Chou's Chaos Game Feature Extraction method
}
