% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIIP.R
\name{EIIP}
\alias{EIIP}
\title{EIIP}
\usage{
EIIP(Seq, EIIP = c(0.126, 0.1335, 0.0806, 0.134))
}
\arguments{
\item{Seq}{A character string which denotes the DNA/RNA sequence}

\item{EIIP}{A numeric vector contains the A/T/G/C EIIP values.
default=A:0.1260, T:0.1335, G:0.0806, C:0.1340}
}
\value{
A vector contains the 3-mer EIIP values
}
\description{
$electron-ion interaction pseudopotentials$ is defined by Nair in 2006.
The EIIP for each nucleutide is $A:0.1260, T:0.1335, G:0.0806, C:0.1340$
}
\details{
The function implement the 3mer EIIP values. EIIP values for each 3-mer is defined as the
mean nuclueotide EIIPs. Example: ATC' EIIP is \emph{average(A's EIIP, T's EIIP, C's EIIP) = (0.126+
0.1335+0.1340)/3}
}
