% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fscore.R
\name{Fscore_each}
\alias{Fscore_each}
\title{Fscore}
\usage{
Fscore_each(feature, Pos)
}
\arguments{
\item{Pos}{A numeric vector contains the positive index.}

\item{features}{A numeric matrix/data.frame which denotes features. Each row is a sample}
}
\value{
A number which is the fscore
}
\description{
The formula is from doi:10.3389/fmicb.2018.02174. A high f-score means
feature has more powerful discrimination ability.
}
\details{
The function implement the F-score method to select the features. It returns a
}
