% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseKNC.R
\name{getDNAClass}
\alias{getDNAClass}
\title{getDNAClass}
\usage{
getDNAClass(nuc)
}
\arguments{
\item{nuc}{A character belongs to A/T/C/G}
}
\value{
A vector contains the 3 numeric values, only 0 or 1
}
\description{
The function return 3 numeric values, denoted theRing struc, Functional group and
Hydrogen bonding
}
