% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{methods}
\name{getEIIP}
\alias{getEIIP}
\title{getEIIP}
\usage{
getEIIP(object, EIIP = c(0.126, 0.1335, 0.0806, 0.134))
}
\arguments{
\item{object}{A Benneat object}
}
\value{
A matrix contains ALL sequences' EIIP features,each row denotes
a sequence. The Sequence is sorted by its order in Benneat object.
}
\description{
getEIIP
}
