% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{methods}
\name{getNCC}
\alias{getNCC}
\title{getNCC}
\usage{
getNCC(object, seqType, n)
}
\arguments{
\item{object}{A Benneat object}

\item{seqType}{The sequence is DNA or RNA. It should be either 'DNA' or 'RNA'}

\item{n}{The k of k-mer}
}
\value{
A matrix contains ALL sequences' NCC features,each row denotes
a sequence. The Sequence is sorted by its order in Benneat object. if you run
getNCC(xxx,"DNA",2), then each col will be sorted according to AA,AC,AG,AT...
TA,TC,TG,TT.
}
\description{
getNCC
}
