% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{methods}
\name{getPseKNC}
\alias{getPseKNC}
\title{getPseKNC}
\usage{
getPseKNC(object)
}
\arguments{
\item{object}{A Benneat object}
}
\value{
A matrix contains ALL sequences' PseKNC features,each row denotes
a sequence. The Sequence is sorted by its order in Benneat object.
}
\description{
getPseKNC
}
