% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{methods}
\name{getT2PseKNC}
\alias{getT2PseKNC}
\title{getT2PseKNC}
\usage{
getT2PseKNC(object, phychem_file = NULL, normalization = T,
  lambda = 4)
}
\arguments{
\item{object}{A Benneat object}

\item{phychem_file}{A matrix;whose rownames are features and colnames are 2-mer}

\item{normalization}{Bool, if TRUE, the function will perform normalization. default:T}

\item{lambda}{numeric. The max number of sequence tiers. default:4}
}
\value{
A matrix contains ALL sequences' moditified T2PseKNC features,each row denotes
a sequence. The Sequence is sorted by its order in Benneat object. The return
value only contains the long-range part of T2PSeKNC. The detailed explanation
can be find in Hao Lin Groups' paper iTerm-PseKNC. This function only implement
the 2-mer situations. The ncol of result will be as lambda*(phechem_file feature nums)-length
}
\description{
Please make sure the sequence type of your data. The default file is 6 features of double-strand B-DNA
}
