% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{method}
\name{makeBenneat}
\alias{makeBenneat}
\title{makeBenneat}
\usage{
makeBenneat(Seqs, labs)
}
\arguments{
\item{Seqs}{A vector contains the sequences}

\item{labs}{A vector contains the labels like "methylated"/"unmethylated"}
}
\value{
A Benneat Class
}
\description{
Make Benneat Class by TWO parameters
}
