% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseKNC.R
\name{PseKNC_each}
\alias{PseKNC_each}
\title{PseKNC}
\usage{
PseKNC_each(i, aa, seq)
}
\value{
A vector contains the PseKNC's nucleotide part, the length is same as the
        input sequence
}
\description{
The function returns a vector which return the PseKNC's nucleatiode part
PseKNC is the daughter-function for the funtion which is designed to lapply
}
