% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{method}
\name{PSTNPds}
\alias{PSTNPds}
\title{PSTNPds}
\usage{
PSTNPds(object, lableA, lableB)
}
\value{
A matrix contains ALL sequences' PSTNPss features,each row denotes
a sequence. The Sequence is sorted by its order in Benneat object. Columns
are sorted by AAA,AAC,ACA,ACC...CCC <8 columns in total>.
}
\description{
Feature PSTNPds
using a statistical strategy based on
double-stranded characteristics of DNA according to
complementary base pairing, so they have more evident
statistical features. At this point, we deem A and T as
identical, the same to C and G. Thus, for every sample,
it can be converted into a sequence contained A and T
only.
}
\section{Slots}{

\describe{
\item{\code{object}}{A Benneat object}

\item{\code{lableA}}{character,Label as the POSITIVE label}

\item{\code{lableB}}{character,Label as the POSITIVE label}
}}

