% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{methods}
\name{PSTNPss}
\alias{PSTNPss}
\title{PSTNPss}
\usage{
PSTNPss(object, lableA, lableB)
}
\arguments{
\item{object}{A Benneat object}

\item{lableA}{character,Label as the POSITIVE label}

\item{lableB}{character,Label as the POSITIVE label}
}
\value{
A matrix contains ALL sequences' PSTNPss features,each row denotes
a sequence. The Sequence is sorted by its order in Benneat object. Each col
is sorted by AAA,AAC,AAG............TTT
}
\description{
PSTNPss
}
