% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCC.R
\name{NCC}
\alias{NCC}
\title{NCC}
\usage{
NCC(Seq, seqType, n)
}
\arguments{
\item{Seq}{A character string which denotes the DNA/RNA sequence}

\item{seqType}{The sequence is DNA or RNA. It should be either 'DNA' or 'RNA'}

\item{n}{The k of k-mer}
}
\value{
A vector contains the compositions
}
\description{
The output is sorted by A/T/G/C at each posotions. For example, if you run
NCC(xxx,"DNA",2), then you'll get a vector which is sorted according to AA,AT,AG,AC,TA,TT,TG,
TC,GA,GT,GG,.....CC.
}
\details{
The function implement the basic DNA/RNA Sequence nucleotide composition counting.
}
