% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benneat.R
\docType{method}
\name{PSTNPss}
\alias{PSTNPss}
\title{PSTNPss}
\usage{
PSTNPss(object, lableA, lableB)
}
\value{
A matrix contains ALL sequences' PSTNPss features,each row denotes
a sequence. The Sequence is sorted by its order in Benneat object
}
\description{
PSTNPss
}
\section{Slots}{

\describe{
\item{\code{object}}{A Benneat object}

\item{\code{lableA}}{character,Label as the POSITIVE label}

\item{\code{lableB}}{character,Label as the POSITIVE label}
}}

